Renamed pauli_container to pauli_term_container
