+++ include/observable.hpp
@@ -3,8 +3,7 @@
 
 #include "pauli.hpp"
 #include "pauli_term.hpp"
-#include "pauli_container.hpp"
-#include "non_owning_pauli_term.hpp"
+#include "pauli_term_container.hpp"
 #include "merge.hpp"
 
 #include <algorithm>
